issue #226, type change caused error on mac, that has been fixed

+++ odmtools/odmservices/edit_service.py
@@ -304,7 +304,7 @@ class EditService():
     def delete_points(self):
         filtered_points = self.get_filtered_points()
         if not filtered_points.empty:
-            values = filtered_points['ValueID'].astype(np.int32).tolist()
+            values = filtered_points['ValueID'].astype(float).tolist()
 
             self.memDB.delete(values)
             self._populate_series()
